Extract rendering into separate file.
