+++ lib/cuffsert/main.rb
@@ -5,6 +5,7 @@ require 'cuffsert/confirmation'
 require 'cuffsert/messages'
 require 'cuffsert/metadata'
 require 'cuffsert/presenters'
+require 'cuffsert/rendering'
 require 'cuffsert/rxcfclient'
 require 'cuffsert/rxs3client'
 require 'rx'
